refactor(coordinator): extract registry and orchestrator
